Updated PLibOptions 3.8 for Cloud Save update

+++ MiniBase/MiniBasePatches.cs
@@ -83,7 +83,7 @@ namespace MiniBase
                 if (!IsMiniBase())
                     return;
                 Log("MinionSelectScreen_OnProceed_Patch Postfix");
-                float radius = Math.Max(Grid.WidthInCells, Grid.HeightInCells) * 1.5f;
+                int radius = (int) (Math.Max(Grid.WidthInCells, Grid.HeightInCells) * 1.5f);
                 GridVisibility.Reveal(0, 0, radius, radius - 1);
             }
         }
